--- a/Common/Securities/Positions/OptionStrategyPositionGroupResolver.cs
+++ b/Common/Securities/Positions/OptionStrategyPositionGroupResolver.cs
@@ -59,10 +59,25 @@ namespace QuantConnect.Securities.Positions
         /// <returns>True if this resolver can group the specified positions, otherwise false</returns>
         public bool TryGroup(IReadOnlyCollection<IPosition> newPositions, PositionGroupCollection currentPositions, out IPositionGroup @group)
         {
-            var impactedGroups = GetImpactedGroups(currentPositions, newPositions);
-            var positionsToConsiderInNewGroup = impactedGroups.SelectMany(positionGroup => positionGroup.Positions);
+            IEnumerable<IPosition> positions;
+            if (currentPositions.Count > 0)
+            {
+                var impactedGroups = GetImpactedGroups(currentPositions, newPositions);
+                var positionsToConsiderInNewGroup = impactedGroups.SelectMany(positionGroup => positionGroup.Positions);
+                positions = newPositions.Concat(positionsToConsiderInNewGroup);
+            }
+            else
+            {
+                if (newPositions.Count == 1)
+                {
+                    // there's no existing position and there's only a single position, no strategy will match
+                    @group = null;
+                    return false;
+                }
+                positions = newPositions;
+            }
 
-            @group = GetPositionGroups(newPositions.Concat(positionsToConsiderInNewGroup)).Where(positionGroup =>
+            @group = GetPositionGroups(positions).Where(positionGroup =>
             {
                 // from the resolved position groups we will take those which use our buying power model and which are related to the new positions to be executed
                 if (positionGroup.BuyingPowerModel.GetType() == typeof(OptionStrategyPositionGroupBuyingPowerModel))
@@ -109,6 +124,12 @@ namespace QuantConnect.Securities.Positions
         /// <returns>An enumerable containing the position groups that could be impacted by the specified position changes</returns>
         public IEnumerable<IPositionGroup> GetImpactedGroups(PositionGroupCollection groups, IReadOnlyCollection<IPosition> positions)
         {
+            if(groups.Count == 0)
+            {
+                // there's no existing groups, nothing to impact
+                return Enumerable.Empty<IPositionGroup>();
+            }
+
             var symbolsSet = positions.Where(position => position.Symbol.SecurityType.HasOptions() || position.Symbol.SecurityType.IsOption())
                 .SelectMany(position =>
                 {