… Common/Securities/Positions/PositionGroupKey.cs
… Common/Securities/SecurityHolding.cs
… Common/Securities/SecurityManager.cs
… Common/Securities/SecurityPortfolioManager.cs
… Engine/AlgorithmManager.cs
… Engine/DataFeeds/DataManager.cs
… Tests/Algorithm/Framework/Portfolio/PortfolioTargetCollectionTests.cs
… Tests/Common/Securities/Positions/PositionGroupCollectionTests.cs