<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Remove Immutable collection from position groups to improve performance. Adding more tests
- Remove ConcurrentDictionary from PortfolioTargetCollection & DataManager. Adding more tests
- Add Securities enumerator keys & values cache

```csharp
    Initialize()
    {
        SetStartDate(2015, 9, 1);
        SetEndDate(2016, 6, 1);
        foreach (var symbol in Symbols.Equity.All.Take(400))
        {
            AddSecurity(SecurityType.Equity, symbol);
        }
        SetExecution(new ImmediateExecutionModel());
        SetRiskManagement(new MaximumDrawdownPercentPerSecurity(0.01m));
        SetPortfolioConstruction(new EqualWeightingPortfolioConstructionModel(Resolution.Daily));
        SetAlpha(new ConstantAlphaModel(InsightType.Price, InsightDirection.Up, TimeSpan.FromMinutes(10), 0.025, null));
    }
```

`MASTER @2.5GHz`
29.68 seconds at 58k data points per second. Processing total of 1,731,464 data points.
29.90 seconds at 58k data points per second. Processing total of 1,731,464 data points.
29.87 seconds at 58k data points per second. Processing total of 1,731,464 data points.

`PR @2.5GHz`
24.52 seconds at 71k data points per second. Processing total of 1,731,464 data points.
24.76 seconds at 70k data points per second. Processing total of 1,731,464 data points.
24.94 seconds at 69k data points per second. Processing total of 1,731,464 data points.

#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
N/A
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Speed
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Existing tests, sanity tested C#/Py backtests, live & research
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [ ] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->